# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bob - Entrepreneur Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses/{id}/tasks:
    get:
      operationId: listBusinessTasks
      tags: [tasks]
      summary: List tasks for a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task for a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task (status, details, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, closed, waiting_approval]
        assignedAgent:
          type: ["string", "null"]
        agentType:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high, critical]
        estimatedHours:
          type: ["number", "null"]
        deliverables:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        lastProgressUpdate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - businessId
        - title
        - description
        - status
        - assignedAgent
        - agentType
        - priority
        - estimatedHours
        - deliverables
        - dueDate
        - createdAt
        - updatedAt
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, closed, waiting_approval]
        assignedAgent:
          type: ["string", "null"]
        agentType:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high, critical]
        estimatedHours:
          type: ["number", "null"]
        deliverables:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
      required:
        - title
        - description
        - priority
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, closed, waiting_approval]
        assignedAgent:
          type: ["string", "null"]
        agentType:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high, critical]
        estimatedHours:
          type: ["number", "null"]
        deliverables:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
...