# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bob - Entrepreneur Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskPriority } from "./taskPriority";
import type { TaskStatus } from "./taskStatus";

export interface Task {
  id: number;
  businessId: number;
  title: string;
  description: string;
  status: TaskStatus;
  /** @nullable */
  assignedAgent: string | null;
  /** @nullable */
  agentType: string | null;
  priority: TaskPriority;
  /** @nullable */
  estimatedHours: number | null;
  /** @nullable */
  deliverables: string | null;
  /** @nullable */
  dueDate: string | null;
  /** @nullable */
  lastProgressUpdate?: string | null;
  createdAt: string;
  updatedAt: string;
}