# lib/api-zod/src/generated/types/updateTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bob - Entrepreneur Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTaskBodyPriority } from "./updateTaskBodyPriority";
import type { UpdateTaskBodyStatus } from "./updateTaskBodyStatus";

export interface UpdateTaskBody {
  title?: string;
  description?: string;
  status?: UpdateTaskBodyStatus;
  /** @nullable */
  assignedAgent?: string | null;
  /** @nullable */
  agentType?: string | null;
  priority?: UpdateTaskBodyPriority;
  /** @nullable */
  estimatedHours?: number | null;
  /** @nullable */
  deliverables?: string | null;
  /** @nullable */
  dueDate?: string | null;
}